Split direct build host effect tests
