+++ docs/PHASE_PLAN.md
@@ -515,8 +515,9 @@ checked-in docs, tests, and commits only.
   multi-target, and dependency-order compilation coverage in a focused child
   module, leaving CLI usage and legacy JSON diagnostics in the parent module.
 - Direct `zen build.zen` host-effect tests now keep declared file-read fallback
-  accept/reject cases in a focused child module, leaving env-effect ordering and
-  unrelated test skip coverage in the parent integration module.
+  accept/reject cases in a focused child module and env-read coverage in
+  `tests/integration/cli_build/direct_build_graph_host_effects/env_reads/`,
+  leaving the parent module as wiring only.
 - Emit-command build graph host-effect tests now keep declared file-read
   fallback accept/reject cases in a focused child module, leaving env-effect
   ordering and unrelated test skip coverage in the parent integration module.
